requirements.txt
weasyprint
flask
Flask-SocketIO==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0
eventlet
gunicorn
flask_cors
requests
python-nmap
pytz
